requirements.txt

pandas
numpy
scikit-learn>=1.3.0
sentence-transformers>=2.6.1
transformers>=4.41.0
torch>=2.2.0
nltk
spacy
langdetect
datasketch
streamlit
shap
matplotlib
python-dotenv
kaggle
tqdm
rich
Pillow
requests
pyyaml
